Switch from Webpack & Babel to Rollup & Sucrase.
This increases build times quite a lot and the config
is much, much cleaner. Some hacks are still needed,
see packages/app/rollup.config.js.

This also adds a much better way to handle extensions. Extensions are
listed/imported in ./extensions.js. This file (should be ES5) is
imported by both the frontend and the backend - whereas rollup
(frontend) looks for the "browser" key in package.json as its
entrypoint, node (backend) looks for the "main" key. See packages/graph
for an example.